Add files via upload
We are provided with hourly rental data for two years. For this competition, the training set consists of the first 19 days of each month, and the test set consists of the 20th to the end of the month. We must predict the total number of bikes rented during each hour covered by the test set, using only the information available before the start of the rental period.